Add support for plugins, update documentation, bump version to v3.0
MCDeploy can now download plugins, in a similar way to the datapack
download mechanism. You still need to override the Fetcher in the config
in order to use a server.jar that supports loading plugins.

Documentation has been updated to better conform to the actual behavior.

Datapacks and plugins can have their download locations overridden,
using the TargetDir option in the respective section.

The default config has been moved into the resources folder instead of
staying as a wall of text in the WallOfText.kt file. Due to that,
there's no need for that file anymore.

Since this is a breaking change (and a pretty major one) we release a
new version, v3.0!

Signed-off-by: Kozova1 <mug66kk@gmail.com>